Avoid contacting AUR if state = present
When checking whether a package needs to be installed, if we find the
package does not exist locally then there is no need to contact the AUR
server to see what version is available; the knowledge that the package
is not already installed should be enough. This makes subsequent Ansible
runs faster.

We'll still hit an error when we try to install a package that does not
exist.